refactor(follow): 패키지명 변경 entity -> model
#7
